Periodic_Updates_25.08.25
1. 방 배치 배치도 추가
2. 적을 처치 시 아이템 드랍 세분화
3. 미니맵에 현재 위치가 뜨지않았던 (UI) 버그 해결
4. 비율에 따라 확률이 달라지는 시스템을 Float에서 Int형으로 사용하여 최적화

+++ Assets/Scripts/Controller/SortingObject/InteractableBuild/MainGame/Destructible/Prison/PrisonController.cs
@@ -60,8 +60,8 @@ public class PrisonController : InteractableBuildController
     [HideInInspector] protected List<SpriteRenderer> PrisonAllyShadowList = new List<SpriteRenderer>();
     [HideInInspector] protected PrisonAllySprite AllySprites;
 
-    [HideInInspector] private static List<float> PercentPrisonGrade = new List<float>
-    { 5f, 4f, 3f, 2f, 1f };
+    [HideInInspector] private static List<int> PercentPrisonGrade = new List<int>
+    { 5, 4, 3, 2, 1 };
 
     #endregion
